third_party/blink/web_tests/accessibility/continuation.html

<!DOCTYPE HTML>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>

<b>
  <i>
    <div>
      <button id="before">Before</button>
    </div></i><div id="after">After</div></b>

<script>
test(function(t) {
    var axBefore = accessibilityController.accessibleElementById("before");
    assert_equals(axBefore.name, "Before");
    var axAfter = accessibilityController.accessibleElementById("after");
    assert_equals(axAfter.childAtIndex(0).name, "After");
}, "Ensure that continuations are included in the accessibility tree.");
</script>
